Fix -Wc++11-narrowing on 32-bit machines
Fixes https://github.com/rui314/mold/issues/1127

+++ elf/gdb-index.cc
@@ -639,7 +639,7 @@ template <typename E>
 std::span<u8> get_buffer(Context<E> &ctx, Chunk<E> *chunk) {
   if (chunk->is_compressed)
     return chunk->uncompressed_data;
-  return {ctx.buf + chunk->shdr.sh_offset, chunk->shdr.sh_size};
+  return {ctx.buf + chunk->shdr.sh_offset, (size_t)chunk->shdr.sh_size};
 }
 
 template <typename E>
